+++ tests/test_closest_package_tag.py
@@ -9,10 +9,7 @@ def handler():
     mock_config = Mock()
     mock_config.config = {"arch-repositories": []}
 
-    with (
-        patch("archlog.package_handler.WebScraper"),
-        patch("archlog.package_handler.GitLabAPI"),
-    ):
+    with (patch("archlog.package_handler.WebScraper"),):
         return PackageHandler(mock_logger, mock_config)
 
 
@@ -21,9 +18,19 @@ def test_exact_match(handler):
     assert handler.get_closest_package_tag("48.0", tags) == "48.0"
 
 
-def test_close_match(handler):
+def test_close_match_suffix_1(handler):
     tags = ["48.0", "48.1", "48.alpha", "48.beta"]
     assert handler.get_closest_package_tag("48.0-1", tags) == "48.0"
+
+
+def test_close_match_suffix_2(handler):
+    tags = ["20250716", "20250707", "20250430.1", "20250430", "20250123.1"]
+    assert handler.get_closest_package_tag("20250707-1", tags) == "20250707"
+
+
+def test_close_match_prefix_suffix(handler):
+    tags = ["v6.3.1", "v6.3.0", "v6.2.90", "v6.2.91", "v6.3.90", "v6.3.91"]
+    assert handler.get_closest_package_tag("1-6.3.90-1", tags) == "v6.3.90"
 
 
 def test_close_match_release_candidate(handler):