+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - **Web Scraper** Remove all references regarding the creation date of package tags since it is currently not relevant and not used
 - **General** Change changelog filename from "%Y-%m-%d-changelog.json" to "%Y%m%d-%H%M-changelog.json" in order to create multiple changelog files per day
 - **GitHub API** Initial version of the GitHub REST API implementation
+- **Package Handler**, **Test** Extend regex of get_closest_package_tag and tests to handle additional cases
 
 ### Bug fixes